tests

+++ main/src/server/domain/users/users.service.ts
@@ -1,15 +1,17 @@
 import {
-  UserDto,
   validateCreate,
-  toUser,
-  toUserUpdate,
-  validateUpdate,
-} from '../types';
+  toPrismaUserCreate,
+  validateSelfUpdate,
+  FrontendUserCreateDto,
+  FrontendSelfUpdateDto,
+  FrontendUpdateOtherUserDto,
+} from './types';
 import { UserViewModel } from './../auth/auth.service';
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
-  UnauthorizedException,
+  NotFoundException,
 } from '@nestjs/common';
 import { PrismaService } from '../../prisma-connect/prisma.service';
 import { Permission, User, UserPermissions } from '@prisma/client';
@@ -40,12 +42,15 @@ export class UsersService {
       .then((users: User[]) => users.map(this.toViewModel));
   }
 
-  async create(user: LoggedInUser, userDatta: UserDto): Promise<UserViewModel> {
-    if (!validateCreate(userDatta)) {
+  async create(
+    user: LoggedInUser,
+    userData: FrontendUserCreateDto,
+  ): Promise<UserViewModel> {
+    if (!validateCreate(userData)) {
       throw new BadRequestException();
     }
     const createdUser = await this.prisma.user.create({
-      data: await toUser(userDatta, this.bcryptService.hashData),
+      data: await toPrismaUserCreate(userData, this.bcryptService.hashData),
     });
     await this.logsService.log({
       message: `dodał użytkownika ${createdUser.firstName} ${createdUser.lastName} (${createdUser.login})`,
@@ -55,38 +60,60 @@ export class UsersService {
     return this.toViewModel(createdUser);
   }
 
-  async update(
-    id: number,
-    user: UserDto,
+  async updateSelf(
+    user: FrontendSelfUpdateDto,
     loggedInUser: LoggedInUser,
-  ): Promise<UserViewModel> {
-    if (
-      !loggedInUser.permissions.includes(Permission.USER) &&
-      id !== loggedInUser.id
-    ) {
-      throw new UnauthorizedException();
-    }
-    if (!validateUpdate(id, user)) {
+  ): Promise<FrontendSelfUpdateDto> {
+    if (!validateSelfUpdate(user)) {
       throw new BadRequestException();
     }
-    const prevUser = await this.findById(user.id);
-    if (prevUser?.id !== id) {
-      throw new BadRequestException(id, 'id musi się zgadzać');
+    const prevUser = await this.findById(loggedInUser.id);
+    if (!prevUser) {
+      // shouldn't happpen, but...
+      throw new NotFoundException();
     }
-    const prevPermissions = await this.getPermissions(id);
 
-    await this.prisma.userPermissions.deleteMany({ where: { userId: id } });
-
-    const data = toUserUpdate(user);
-    const updatedUser = await this.prisma.user.update({ where: { id }, data });
+    const updatedUser = await this.prisma.user.update({
+      where: { id: loggedInUser.id },
+      data: user,
+    });
 
     // add passwordHash to satisfy the types
-    const diff = formattedDiff(prevUser, { ...user, passwordHash: null }, [
+    const prevUserDiffData = {
+      login: prevUser.login,
+      firstName: prevUser.firstName,
+      lastName: prevUser.lastName,
+    };
+    const diff = formattedDiff(prevUserDiffData, user, [
       { name: 'Login', selector: (n: User) => n.login },
       { name: 'Imię', selector: (n: User) => n.firstName },
       { name: 'Nazwisko', selector: (n: User) => n.lastName },
-      { name: 'Aktywność', selector: (n: User) => n.isActive },
     ]);
+
+    await this.logsService.log({
+      message: `zaktualizował użytkownika ${prevUser.login} (id: ${prevUser.id}) Dane: ${diff}`,
+      permission: Permission.USER,
+      user: loggedInUser,
+    });
+    return this.toViewModel(updatedUser);
+  }
+
+  async updateOther(
+    id: number,
+    body: FrontendUpdateOtherUserDto,
+    loggedInUser: LoggedInUser,
+  ) {
+    if (loggedInUser.id === id) {
+      throw new ForbiddenException();
+    }
+
+    const prevUser = await this.findById(id);
+    if (!prevUser) {
+      throw new NotFoundException();
+    }
+
+    const prevPermissions = await this.getPermissions(id);
+    await this.prisma.userPermissions.deleteMany({ where: { userId: id } });
 
     const permissionNames = {
       USER: 'Użytkownicy',
@@ -97,18 +124,39 @@ export class UsersService {
 
     const permissionsDiff = havePermissionsChanged(
       prevPermissions.map((p) => p.permission),
-      user.permissions,
+      body.permissions,
     )
-      ? ` (Zmienione uprawnienia: ${user.permissions.map(
+      ? ` (Zmienione uprawnienia: ${body.permissions.map(
           (p) => permissionNames[p],
         )})`
       : '';
+
+    const diff = formattedDiff(
+      { isActive: prevUser.isActive },
+      { isActive: body.isActive },
+      [
+        {
+          name: 'Aktywność',
+          selector: (n: { isActive: boolean }) => n.isActive,
+        },
+      ],
+    );
+
+    const updatedUser = await this.prisma.user.update({
+      where: { id },
+      data: { isActive: body.isActive },
+    });
+
+    await this.prisma.userPermissions.createMany({
+      data: body.permissions.map((p) => ({ permission: p, userId: id })),
+    });
 
     await this.logsService.log({
       message: `zaktualizował użytkownika ${prevUser.login} (id: ${prevUser.id}) Dane: ${diff}${permissionsDiff}`,
       permission: Permission.USER,
       user: loggedInUser,
     });
+
     return this.toViewModel(updatedUser);
   }
 
@@ -116,6 +164,10 @@ export class UsersService {
     loggedInUser: LoggedInUser,
     userId: number,
   ): Promise<UserViewModel> {
+    if (loggedInUser.id === userId) {
+      throw new BadRequestException();
+    }
+
     await this.prisma.userPermissions.deleteMany({ where: { userId } });
 
     const deletedUser = await this.prisma.user.delete({
